fix bucket delete

+++ development/tools/cmd/gcscleaner/main.go
@@ -5,7 +5,10 @@ import (
 	"flag"
 	"regexp"
 	"strings"
+	"sync"
 	"time"
+
+	"google.golang.org/api/iterator"
 
 	"cloud.google.com/go/storage"
 	"github.com/kyma-project/test-infra/development/tools/pkg/gcscleaner"
@@ -15,14 +18,17 @@ import (
 
 // Config structure aggregating application configuration arguments
 type options struct {
-	ProjectName            string
-	BucketLifespanDuration time.Duration
-	ExcludedBucketNames    []string
-	DryRun                 bool
-	BucketNameRegexp       regexp.Regexp
+	ProjectName               string
+	BucketLifespanDuration    time.Duration
+	ExcludedBucketNames       []string
+	DryRun                    bool
+	BucketNameRegexp          regexp.Regexp
+	BucketObjectWorkersNumber int
 }
 
 func main() {
+
+	logrus.SetLevel(logrus.DebugLevel)
 	ctx := context.Background()
 
 	options, err := readOptions()
@@ -40,6 +46,74 @@ func main() {
 		deleteBucket = dryRunDeleteBucket
 	} else {
 		deleteBucket = func(bucketName string) error {
+			ctx, cancel := context.WithCancel(context.Background())
+			bucketNamesChan := make(chan string)
+			errorChan := make(chan error)
+			go func() {
+				defer close(bucketNamesChan)
+				objects := client.Bucket(bucketName).Objects(ctx, nil)
+				for {
+					select {
+					case <-ctx.Done():
+						return
+					default:
+						objectAttrs, err := objects.Next()
+						if err == iterator.Done {
+							return
+						}
+						if err != nil {
+							logrus.Error(err)
+							errorChan <- err
+							cancel()
+						}
+						if objectAttrs == nil {
+							logrus.Error(ErrInvalidBucketObjectAttributes)
+							errorChan <- err
+							cancel()
+						}
+						bucketNamesChan <- objectAttrs.Name
+					}
+				}
+			}()
+
+			var waitGroup sync.WaitGroup
+			waitGroup.Add(options.BucketObjectWorkersNumber)
+			go func() {
+				defer close(errorChan)
+				for i := 0; i < options.BucketObjectWorkersNumber; i++ {
+					go func() {
+						defer waitGroup.Done()
+						for {
+							select {
+							case <-ctx.Done():
+								return
+							default:
+							}
+							select {
+							case objectName, ok := <-bucketNamesChan:
+								if !ok {
+									return
+								}
+								logrus.Info("deleting ", bucketName, "::", objectName)
+								if err := client.Bucket(bucketName).Object(objectName).Delete(ctx); err != nil {
+									logrus.Error(err)
+									errorChan <- err
+									cancel()
+								}
+							default:
+							}
+						}
+					}()
+				}
+			}()
+			waitGroup.Wait()
+			var allErrors []string
+			for err := range errorChan {
+				allErrors = append(allErrors, err.Error())
+			}
+			if len(allErrors) > 0 {
+				return errors.Wrap(ErrDeletingBucketObjects, strings.Join(allErrors, "\n"))
+			}
 			return client.Bucket(bucketName).Delete(ctx)
 		}
 	}
@@ -67,13 +141,15 @@ func main() {
 }
 
 var (
-	argProjectName                string
-	argExcludedBucketNames        string
-	argBucketLifespanDuration     string
-	argBucketNameRegexp           string
-	argDryRun                     bool
-	bucketLifespanDurationDefault = "2h"
-	dryRunDeleteBucket            = func(_ string) error {
+	argProjectName                  string
+	argExcludedBucketNames          string
+	argBucketLifespanDuration       string
+	argBucketNameRegexp             string
+	argDryRun                       bool
+	argBucketObjectWorkerNumber     int
+	bucketLifespanDurationDefault   = "2h"
+	bucketObjectWorkerNumberDefault = 1
+	dryRunDeleteBucket              = func(_ string) error {
 		return nil
 	}
 	// ErrInvalidProjectName returned if project name argument is invalid
@@ -84,6 +160,10 @@ var (
 	ErrEmptyBucketNameRegexp = errors.New("empty bucketNameRegexp argument")
 	// ErrInvalidBucketNameRegexp returned if argBucketNameRegexp is invalid
 	ErrInvalidBucketNameRegexp = errors.New("invalid bucketNameRegexp argument")
+	// ErrDeletingBucketObjects returned when bucket object deletion was unsuccessful
+	ErrDeletingBucketObjects = errors.New("error while deleting bucket object")
+	// ErrInvalidBucketObjectAttributes returned when bucket object has invalid attributes
+	ErrInvalidBucketObjectAttributes = errors.New("bucket object has invalid attributes")
 )
 
 func init() {
@@ -117,6 +197,12 @@ func init() {
 		"bucketNameRegexp",
 		"",
 		"bucket name regexp pattern used to mach when deleted buckets")
+
+	flag.IntVar(
+		&argBucketObjectWorkerNumber,
+		"bucketObjectWorkerNumber",
+		bucketObjectWorkerNumberDefault,
+		"the number of workers that will be used to delete bucket object")
 }
 
 func readOptions() (options, error) {
@@ -150,6 +236,6 @@ func readOptions() (options, error) {
 	}
 	options.DryRun = argDryRun
 	options.BucketNameRegexp = *bucketNameRegexp
-
+	options.BucketObjectWorkersNumber = argBucketObjectWorkerNumber
 	return options, nil
 }
